feat: optimize database indexes & implement static array seeders
- Reduced database indexes from 36 to 19 (47% reduction) across 4 tables
- Optimized anggota, keuangan, open_recruitment, and anggaran tables
- Converted seeders to static PHP arrays (AnggotaSeeder: 154 records, OpenRecruitmentSeeder: 199 records)
- Fixed jenis_oprec classification logic (45 pengurus, 153 anggota)
- Redesigned modal detail pages with minimalist Canva-style design
- Added INDEX_OPTIMIZATION.md and IMPORT_ANGGOTA_OPREC_GUIDE.md documentation
- Fixed text overflow issues in anggota and open recruitment modals
- Added Divisi and TentangKami Livewire components
- Updated UI with improved styling and FontAwesome Pro icons

+++ database/migrations/2026_02_07_234645_create_bank_soal_pertemuan_table.php
@@ -31,11 +31,11 @@ return new class extends Migration
             $table->integer('tampil_esai')->default(0);
             
             // Weights per type
-            $table->decimal('bobot_pg', 8, 2)->default(0);
-            $table->decimal('bobot_kompleks', 8, 2)->default(0);
-            $table->decimal('bobot_jodohkan', 8, 2)->default(0);
-            $table->decimal('bobot_isian', 8, 2)->default(0);
-            $table->decimal('bobot_esai', 8, 2)->default(0);
+            $table->integer('bobot_pg')->default(0);
+            $table->integer('bobot_kompleks')->default(0);
+            $table->integer('bobot_jodohkan')->default(0);
+            $table->integer('bobot_isian')->default(0);
+            $table->integer('bobot_esai')->default(0);
             
             // Multiple choice options count (3, 4, or 5)
             $table->enum('opsi', ['3', '4', '5'])->default('5');
